feat: progress bar

+++ src/main/kotlin/io/github/trcdevelopers/clayium/api/capability/IClayiumWorkable.kt
@@ -4,7 +4,7 @@ package io.github.trcdevelopers.clayium.api.capability
  * Capability interface for TileEntities.
  */
 interface IClayiumWorkable {
-    val currentProgress: Int
-    val requiredProgress: Int
+    val currentProgress: Long
+    val requiredProgress: Long
     val isWorking: Boolean
 }
